requirements.txt
streamlit
PyPDF2
scikit-learn
pandas
numpy
python-docx
openai
spacy
google-generativeai
python-dotenv
PyMuPDF